ENT-4601 Do not use lambda `await` implementations
To make it harder to use code from `ServiceHub`, the `await` implementations
that used lambdas have been removed. The theory behind this is that if a
developer wants to access the `ServiceHub` they still can, but will need
to explicitly pass it into their own implementation. Developers __can__
still use the `ServiceHub` but need to go through the effort to do so.

`FlowBackgroundProcess` renamed to `FlowExternalFuture`

`FlowExternalResult` interface made which developers implement with a normal
function which will be ran on the default thread pool. `FlowExternalResult`
is converted into a `FlowExternalFuture` under the hood.
