--- core/src/main/kotlin/net/corda/core/flows/FlowBackgroundProcess.kt
@@ -1,8 +1,0 @@
-package net.corda.core.flows
-
-import net.corda.core.internal.FlowAsyncOperation
-import net.corda.core.node.ServiceHub
-
-interface FlowBackgroundProcess<R : Any> : FlowAsyncOperation<R>
-
-internal abstract class FlowBackgroundProcessImpl<R : Any>(val serviceHub: ServiceHub) : FlowBackgroundProcess<R>
